Create two renderer instances: block and inline
