+++ src/ServiceProvider/MarkdownServiceProvider.php
@@ -14,26 +14,106 @@
 namespace Eventum\ServiceProvider;
 
 use Eventum\Config\Paths;
+use Eventum\Event;
+use Eventum\EventDispatcher\EventManager;
 use Eventum\Markdown;
+use Eventum\Markdown\CommonMark\UserMentionGenerator;
 use HTMLPurifier;
 use HTMLPurifier_HTML5Config;
+use League\CommonMark\CommonMarkConverter;
+use League\CommonMark\Environment;
+use League\CommonMark\Extension\Attributes\AttributesExtension;
+use League\CommonMark\Extension\Autolink\AutolinkExtension;
+use League\CommonMark\Extension\CommonMarkCoreExtension;
+use League\CommonMark\Extension\Footnote\FootnoteExtension;
+use League\CommonMark\Extension\HeadingPermalink\HeadingPermalinkExtension;
+use League\CommonMark\Extension\InlinesOnly\InlinesOnlyExtension;
+use League\CommonMark\Extension\Mention\MentionExtension;
+use League\CommonMark\Extension\Table\TableExtension;
+use League\CommonMark\Extension\TaskList\TaskListExtension;
+use League\CommonMark\MarkdownConverterInterface;
 use Misc;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
+use Setup;
+use Symfony\Component\EventDispatcher\GenericEvent;
 
 class MarkdownServiceProvider implements ServiceProviderInterface
 {
     private const PURIFIER_CACHE_DIR = Paths::APP_CACHE_PATH . '/purifier';
 
+    /**
+     * Use moderately sane value
+     *
+     * @see https://commonmark.thephpleague.com/security/#nesting-level
+     */
+    private const MAX_NESTING_LEVEL = 500;
+
     public function register(Container $app): void
     {
-        $app[Markdown\MarkdownRendererInterface::class] = static function ($app) {
-            return new Markdown\MarkdownRenderer($app[HTMLPurifier::class]);
+        $app[Markdown\MarkdownRendererInterface::RENDER_BLOCK] = function ($app) {
+            return new Markdown\MarkdownRenderer($this->createConverter(false), $app[HTMLPurifier::class]);
+        };
+        $app[Markdown\MarkdownRendererInterface::RENDER_INLINE] = function ($app) {
+            return new Markdown\MarkdownRenderer($this->createConverter(true), $app[HTMLPurifier::class]);
         };
 
         $app[HTMLPurifier::class] = static function () {
             return self::createPurifier();
         };
+    }
+
+    private function createConverter(bool $inline): MarkdownConverterInterface
+    {
+        $config = [
+            'renderer' => [
+                'block_separator' => "\n",
+                'inner_separator' => "\n",
+                'soft_break' => "<br />\n",
+            ],
+            'html_input' => Environment::HTML_INPUT_ALLOW,
+            'allow_unsafe_links' => false,
+            'max_nesting_level' => self::MAX_NESTING_LEVEL,
+            'heading_permalink' => [
+                'inner_contents' => '¶',
+                'insert' => 'after',
+            ],
+
+            // https://commonmark.thephpleague.com/1.5/extensions/mentions/
+            'mentions' => [
+                'eventum_handle' => [
+                    'symbol' => '@',
+                    'regex' => '/^[A-Za-z0-9_]{1,255}(?!\w)/',
+                    'generator' => new UserMentionGenerator(Setup::getBaseUrl()),
+                ],
+            ],
+        ];
+
+        $environment = new Environment($config);
+        if ($inline) {
+            $environment->addExtension(new InlinesOnlyExtension());
+        } else {
+            $environment->addExtension(new CommonMarkCoreExtension());
+        }
+
+        $this->applyExtensions($environment);
+
+        return new CommonMarkConverter([], $environment);
+    }
+
+    private function applyExtensions(Environment $environment): void
+    {
+        $environment->addExtension(new AutolinkExtension());
+        $environment->addExtension(new TaskListExtension());
+        $environment->addExtension(new TableExtension());
+        $environment->addExtension(new HeadingPermalinkExtension());
+        $environment->addExtension(new AttributesExtension());
+        $environment->addExtension(new FootnoteExtension());
+        $environment->addExtension(new MentionExtension());
+
+        // allow extensions to apply behaviour
+        $event = new GenericEvent($environment);
+        EventManager::dispatch(Event\SystemEvents::MARKDOWN_ENVIRONMENT_CONFIGURE, $event);
     }
 
     private static function createPurifier(): HTMLPurifier
